refactor: removeType
