+++ src/main/java/com/wmp/htmlparser/dto/ParseRequestDto.java
@@ -1,27 +1,22 @@
 package com.wmp.htmlparser.dto;
 
-import lombok.*;
-import org.hibernate.validator.constraints.Range;
-
+import com.wmp.htmlparser.util.remove.RemoveType;
 import javax.validation.constraints.Min;
 import javax.validation.constraints.Pattern;
-
-import static com.wmp.htmlparser.util.RemoveType.Constants.TYPE_MAX_NUMBER;
-import static com.wmp.htmlparser.util.RemoveType.Constants.TYPE_MIN_NUMBER;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
 
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
 public class ParseRequestDto {
 
-    @Pattern(
-            regexp = "(http)s?:\\/\\/(www\\.)?[a-zA-Z0-9@:%._\\+~#=]+\\.[a-zA-Z0-9@:%._\\/+-~#=?]+",
-            message = "url 형식이 아닙니다."
-    )
+    @Pattern(regexp = "(http)s?:\\/\\/(www\\.)?[a-zA-Z0-9@:%._\\+~#=]+\\.[a-zA-Z0-9@:%._\\/+-~#=?]+", message = "url 형식이 아닙니다.")
     private String url;
 
-    @Range(min = TYPE_MIN_NUMBER, max = TYPE_MAX_NUMBER, message = "유효하지 않은 removeTypeNumber입니다.")
-    private int removeTypeNumber;
+    private RemoveType removeType;
 
     @Min(value = 1, message = "1이상의 수를 입력해주세요")
     private int outputUnitCount;
