+++ src/main/java/com/wmp/htmlparser/util/remove/RemoveType.java
@@ -1,0 +1,14 @@
+package com.wmp.htmlparser.util.remove;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public enum RemoveType {
+
+    REMOVE_HTML(new RemoveHtml()), REMOVE_NOTHING(new RemoveNothing());
+
+    private final RemoveStrategy removeStrategy;
+
+}
